refactor: organize screens into feature packages
- Restructure screen composables into sub-packages based on their features: `animations_and_gestures`, `bars_and_navigation`, `components`, and `layouts`.
- Add a category field to each navigation item.
- Group the items by category on the main screen for better organization.
- Make the main screen stateless and add a preview.
