+++ app/src/main/java/com/aubynsamuel/androidEssentials/presentation/navigation/Navigation.kt
@@ -6,28 +6,28 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.aubynsamuel.androidEssentials.presentation.screens.BottomSheet
-import com.aubynsamuel.androidEssentials.presentation.screens.BottomSheetScaffold
-import com.aubynsamuel.androidEssentials.presentation.screens.CardSlideAnimationScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.CircularRevealAnimationScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.FABScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.FlipAnimationScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.ListAnimationScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.LoadingAnimation
 import com.aubynsamuel.androidEssentials.presentation.screens.MainScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.MovingGestureScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.PillProgressIndicatorScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.ProgressIndicators
-import com.aubynsamuel.androidEssentials.presentation.screens.PullToRefreshScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.ReflectionAnimationScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.SearchBar
-import com.aubynsamuel.androidEssentials.presentation.screens.ShapesScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.SnackBarScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.SweepLineAnimationScreen
-import com.aubynsamuel.androidEssentials.presentation.screens.TextGradients
-import com.aubynsamuel.androidEssentials.presentation.screens.TopApBarScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.animations_and_gestures.CardSlideAnimationScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.animations_and_gestures.CircularRevealAnimationScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.animations_and_gestures.FlipAnimationScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.animations_and_gestures.ListAnimationScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.animations_and_gestures.LoadingAnimation
+import com.aubynsamuel.androidEssentials.presentation.screens.animations_and_gestures.MovingGestureScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.animations_and_gestures.ReflectionAnimationScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.animations_and_gestures.SweepLineAnimationScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.bars_and_navigation.FABScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.bars_and_navigation.SearchBar
+import com.aubynsamuel.androidEssentials.presentation.screens.bars_and_navigation.SnackBarScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.bars_and_navigation.TopApBarScreen
 import com.aubynsamuel.androidEssentials.presentation.screens.buttons.Buttons
-import com.aubynsamuel.androidEssentials.presentation.screens.carousel.AnimatedCarousel
+import com.aubynsamuel.androidEssentials.presentation.screens.components.PillProgressIndicatorScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.components.ProgressIndicators
+import com.aubynsamuel.androidEssentials.presentation.screens.components.ShapesScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.components.TextGradients
+import com.aubynsamuel.androidEssentials.presentation.screens.layouts.BottomSheet
+import com.aubynsamuel.androidEssentials.presentation.screens.layouts.BottomSheetScaffold
+import com.aubynsamuel.androidEssentials.presentation.screens.layouts.PullToRefreshScreen
+import com.aubynsamuel.androidEssentials.presentation.screens.layouts.carousel.AnimatedCarousel
 
 @RequiresApi(Build.VERSION_CODES.S)
 @Composable
