+++ app/src/main/java/com/aubynsamuel/androidEssentials/presentation/screens/animations_and_gestures/CardSlideAnimationScreen.kt
@@ -1,4 +1,4 @@
-package com.aubynsamuel.androidEssentials.presentation.screens
+package com.aubynsamuel.androidEssentials.presentation.screens.animations_and_gestures
 
 import androidx.compose.animation.Crossfade
 import androidx.compose.animation.core.animateFloatAsState
