+++ app/src/main/java/com/aubynsamuel/androidEssentials/presentation/screens/animations_and_gestures/ReflectionAnimationScreen.kt
@@ -1,4 +1,4 @@
-package com.aubynsamuel.androidEssentials.presentation.screens
+package com.aubynsamuel.androidEssentials.presentation.screens.animations_and_gestures
 
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.animateDpAsState
